requirements.txt
# ============================================================
# Flight Monitor Bot - Dependencias Python
# ============================================================
# Instalar con: pip install -r requirements.txt
# Luego ejecutar: playwright install chromium
# ============================================================

# --- Web Scraping ---
playwright>=1.49.0              # Automatización de navegador (Chromium)
playwright-stealth>=1.0.6       # Anti-detección de bots (oculta Playwright)
curl-cffi>=0.7.0                # HTTP con TLS fingerprint (bypass Cloudflare)
requests>=2.31.0                # HTTP fallback si curl_cffi no está disponible

# --- Base de Datos ---
sqlalchemy>=2.0.25              # ORM para SQLite (modelos, queries)

# --- Notificaciones ---
python-telegram-bot>=20.7       # Bot de Telegram (alertas)

# --- Configuración ---
python-dotenv>=1.0.0            # Carga de variables desde .env

# --- Utilidades ---
aiohttp>=3.11.0                 # HTTP async (para APIs)

# --- Testing ---
pytest>=7.0.0,<8.0.0            # Framework de testing
pytest-asyncio>=0.23.4          # Soporte async para pytest
